requirements.txt
# --- requirements.txt ---
fastapi
uvicorn[standard]
redis
python-dotenv
aiohttp

#benchmark
locust
uhashring

#cache LRU
cachetools